==> Project/code/02_data_visualization.Rmd <==
---
title: "Chapter 1 data visualization"
author: "Connor Quiroz"
date: "2025-02-12"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Data Visualization + Analysis


```{r figure one}
A <- risk_foreign_imports %>%
create_map(fill = "exposure", country.col.name = "consumer_iso3c") +
  labs(fill = "Exposure", tag = "A") + 
  theme(plot.margin = unit(c(0,0,0,0), "cm"),
        plot.tag = element_text(face = "bold"))

B <- risk_foreign_imports %>%
create_map(fill = "sensitivity", country.col.name = "consumer_iso3c")+
  labs(fill = "Sensitivity", tag = "B") + 
  theme(plot.margin = unit(c(0,0,0,0), "cm"),
        plot.tag = element_text(face = "bold"))

C <- risk_foreign_imports %>%
  create_map(fill = "adaptive_capacity", country.col.name = "consumer_iso3c") +
  labs(fill = "Adaptive capacity", tag = "C") + 
  theme(plot.margin = unit(c(0,0,0,0), "cm"),
        plot.tag = element_text(face = "bold"))

comb_1 <- plot_grid(A,B)
(fig_one <- plot_grid(comb_1,C, ncol = 1))

# Save plot
# ggsave("../images/fig_one.jpg", plot = fig_one, device = "jpeg", width = 6, height = 4, units = "in")
```

```{r bivariate color plot}
# remotes::install_github(repo = "lydialucchesi/Vizumap", build_vignettes = TRUE, force = TRUE)
library(knitr)
library(Vizumap)
library(rnaturalearth)
library(sf)
library(dplyr)

# Step 1: Get world shapefile
world_map <- ne_countries(scale = "medium", returnclass = "sf") %>%
  select(iso_a3, geometry)

# Step 2: Merge your risk data to the shapefile
risk_map <- world_map %>%
  left_join(risk, by = c("iso_a3" = "consumer_iso3c"))

# Step 3: Drop geometry to prepare bivariate input
risk_df <- st_drop_geometry(risk_map) %>%
  filter(!is.na(exposure), !is.na(sensitivity))

# Step 4: Create Vizumap bivariate data object
risk_uv <- read.uv(data = risk_df, estimate = "exposure", error = "sensitivity")

# Step 5: Build color palette
risk_palette <- build_palette(
  name = "usr",
  colrange = list(
    colour = c("#ff9d00", "#0238FF"),  # light blue → dark blue
    difC = c(4, 4)
  )
)

# Step 6: Build bivariate map and legend
risk_biv_map <- build_bmap(
  data = risk_uv,
  geoData = risk_map,
  id = "iso_a3",
  palette = risk_palette,
  terciles = TRUE
)

# Build the key first
risk_biv_key <- build_bkey(
  data = risk_uv,
  palette = risk_palette,
  terciles = TRUE
)

attach_key(map = risk_biv_map, mapkey = risk_biv_key)
```


```{r figure two}
# Map vulnerability
(fig_two <- risk_foreign_imports %>%
  create_map(fill = "vulnerability", country.col.name = "consumer_iso3c") +
  labs(fill = "Vulnerability score"))

5# Save plot
ggsave("../images/fig_two.jpg", plot = fig_two, device = "jpeg", width = 4, height = 3, units = "in")
```


```{r figure three}
A <- risk_foreign_imports %>%
  ggplot(aes(x = exposure, y = vulnerability, color = region)) +
  geom_point() +
  theme_cowplot() +
  scale_color_manual(values = artis_palette(6)) +
  labs(x = "Exposure", y = "Risk", color = "", tag = "A") +
  theme(legend.position = "bottom")

# Get legend object
legend <- get_legend(A + theme(
    legend.position = "right",               # default or wherever you like
    legend.key.size = unit(0.4, "cm"),       # shrink legend keys
    legend.text = element_text(size = 8),    # smaller text
    legend.title = element_text(size = 9),   # adjust title size
    legend.spacing.y = unit(0.1, "cm"),      # less vertical spacing between items
    legend.margin = margin(t = 2, r = 2, b = 2, l = 2)
  ))

legend <- ggdraw() +
  draw_grob(legend, x = 0.5, y = 0.5, hjust = 0.5, vjust = 0.5)

# Remove legend
A <- A +
  guides(color = "none")

B <- risk_foreign_imports %>%
  ggplot(aes(x = sensitivity, y = vulnerability, color = region)) +
  geom_point() +
  theme_cowplot() +
  scale_color_manual(values = artis_palette(6)) +
  labs(x = "Sensitivity", y = "Risk", tag = "B") +
  guides(color = "none")



C <- risk_foreign_imports %>%
  ggplot(aes(x = adaptive_capacity, y = vulnerability, color = region)) +
  geom_point() +
  theme_cowplot() +
  scale_color_manual(values = artis_palette(6)) +
  labs(x = "Adaptive capacity", y = "Risk", tag = "C") +
  guides(color = "none", tag = "C")

comb_1 <- plot_grid(A, B, align = "v")
comb_2 <- plot_grid(C,legend)
(fig_three <- plot_grid(comb_1, comb_2, ncol = 1))

# Save plot
ggsave("../images/fig_three.jpg", plot = fig_three, device = "jpeg", width = 6, height = 4, units = "in")
```

```{r}
risk_comparison %>%
  ggplot(aes(x = foreign_contribution_pct)) +
  geom_histogram(color = "black", boundary = 0) +
  theme_light() +
  labs(x = "Foreign import contribution to entire porfolio risk (%)", y = "Number of countries")

risk_comparison %>%
  create_map(fill = "foreign_contribution_pct", country.col.name = "consumer_iso3c")
```

```{r look at consumed invasives}
invasives <- read_parquet("../output/invasives.parquet")
create_map(data = invasives, fill = "invasive", country.col.name = "iso3c", color.scale = FALSE) + scale_fill_manual(values = "darkred", na.value = "darkgrey") +
  labs(fill = "Invasive")
```
